failing test for link decoding (#293)
