+++ git-index/tests/fixtures/make_index/v2_split_index.sh
@@ -1,0 +1,13 @@
+#!/bin/bash
+set -eu -o pipefail
+
+export GIT_INDEX_VERSION=2
+git init -q
+git config commit.gpgsign false
+git config index.threads 1
+
+touch a
+git add a
+git commit -m "empty"
+
+git update-index --split-index
