+++ git-index/tests/index-multi-threaded.rs
@@ -1,1 +1,2 @@
 mod index;
+pub use index::*;
