+++ git-index/tests/index-single-threaded.rs
@@ -1,1 +1,4 @@
+#[cfg(not(feature = "internal-testing-git-features-parallel"))]
 mod index;
+#[cfg(not(feature = "internal-testing-git-features-parallel"))]
+pub use index::*;
